% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlradSpectra.R
\name{AlradSpectra}
\alias{AlradSpectra}
\title{A GUI for preprocessing and modeling soil spectra}
\usage{
AlradSpectra()
}
\description{
AlradSpectra provides a friendly user interface to apply preprocessing functions to soil spectra and run models on your data
}
\author{
Andre Dotto, \email{andrecdot@gmail.com}
}

