% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlradSpectra.R
\name{AlradSpectra}
\alias{AlradSpectra}
\title{A GUI for preprocessing soil spectra and predicting soil properties}
\usage{
AlradSpectra()
}
\description{
AlradSpectra provides a friendly GUI to apply preprocessing functions to soil spectra and predict soil properties using models.
}
\author{
Andre Dotto, \email{andrecdot@gmail.com}
}

