% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlradSpectra.R
\name{AlradSpectra}
\alias{AlradSpectra}
\title{A GUI for preprocessing spectra and predicting environmental variables}
\usage{
AlradSpectra()
}
\description{
AlradSpectra provides a friendly GUI to apply preprocessing functions to spectra and predict environmental variables using models.
}
\author{
Andre C. Dotto, \email{andrecdot@gmail.com}

Diego J. Gris, \email{diegojgris@gmail.com}
}

