% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlradSpectra.R
\name{AlradSpectra}
\alias{AlradSpectra}
\title{A GUI to perform preprocessing, multivariate modeling and prediction using spectroscopic data}
\usage{
AlradSpectra()
}
\description{
AlradSpectra was developed to accomplish tasks such as perform a large range of spectral preprocessing techniques, implement several multivariate calibration methods, which can provide well-fitted and accurate models, statistics assessment, graphical output, validate the models using independent data sets, and predict unknown samples.
}
\author{
Andre C. Dotto, \email{andrecdot@gmail.com}

Diego J. Gris, \email{diegojgris@gmail.com}
}
